src/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.

gunicorn
psycopg2
PPyGIS
PIL
South
django-stdimage
sorl-thumbnail
django-urli18n
django-haystack
xapian-haystack
django-hvad
django-olwidget
